Admin and User panel

+++ app/src/main/java/com/example/coffeeshoponline/model/OrderItem.kt
@@ -6,8 +6,8 @@ data class OrderItem(
     val title: String = "",
     val numberInCart: Int = 0,
     val selectedSize: String = "",
-    val priceSmall: Int = 0,
-    val priceMedium: Int = 0,
-    val priceLarge: Int = 0,
+    val priceSmall: Double = 0.0,
+    val priceMedium: Double = 0.0,
+    val priceLarge: Double = 0.0,
     val picUrl: List<String>? = null
 ) : Serializable
